requirements.txt
torch>=2.0.0
transformers
datasets
huggingface_hub
pyyaml
